+++ logs/k8s_log_streamer.go
@@ -1,4 +1,4 @@
-package service
+package logs
 
 import (
 	"bufio"
@@ -46,7 +46,7 @@ func NewK8sLogStreamer(k8sConfig *rest.Config) (*K8sLogStreamer, error) {
 
 // TailLogs streams log lines for the given LogContext from a Kubernetes pod.
 func (s *K8sLogStreamer) TailLogs(ctx context.Context, logContext *core.LogContext, stream *connect.ServerStream[workflow.TailLogsResponse]) error {
-	pod, container, err := getPrimaryPodAndContainer(logContext)
+	pod, container, err := GetPrimaryPodAndContainer(logContext)
 	if err != nil {
 		return connect.NewError(connect.CodeNotFound, err)
 	}
