+++ runs/service/run_logs_service.go
@@ -13,27 +13,21 @@ import (
 	"github.com/flyteorg/flyte/v2/gen/go/flyteidl2/common"
 	"github.com/flyteorg/flyte/v2/gen/go/flyteidl2/core"
 	"github.com/flyteorg/flyte/v2/gen/go/flyteidl2/workflow"
+	"github.com/flyteorg/flyte/v2/logs"
 	"github.com/flyteorg/flyte/v2/runs/repository/interfaces"
-
-	"github.com/samber/lo"
 )
 
 const defaultMaxConcurrentStreams = 100
 
-// LogStreamer abstracts log fetching from different backends.
-type LogStreamer interface {
-	TailLogs(ctx context.Context, logContext *core.LogContext, stream *connect.ServerStream[workflow.TailLogsResponse]) error
-}
-
 // RunLogsService implements the RunLogsServiceHandler interface.
 type RunLogsService struct {
 	repo     interfaces.Repository
-	streamer LogStreamer
+	streamer logs.LogStreamer
 	sem      *semaphore.Weighted
 }
 
 // NewRunLogsService creates a new RunLogsService.
-func NewRunLogsService(repo interfaces.Repository, streamer LogStreamer) *RunLogsService {
+func NewRunLogsService(repo interfaces.Repository, streamer logs.LogStreamer) *RunLogsService {
 	return &RunLogsService{
 		repo:     repo,
 		streamer: streamer,
@@ -103,27 +97,4 @@ func getLogContextAndClusterForAttempt(ctx context.Context, repo interfaces.Repo
 	}
 
 	return event.GetLogContext(), event.GetCluster(), nil
-}
-
-// getPrimaryPodAndContainer finds the primary pod and container from a LogContext.
-func getPrimaryPodAndContainer(logContext *core.LogContext) (*core.PodLogContext, *core.ContainerContext, error) {
-	if logContext.GetPrimaryPodName() == "" {
-		return nil, nil, fmt.Errorf("primary pod name is empty in log context")
-	}
-
-	pod, found := lo.Find(logContext.GetPods(), func(pod *core.PodLogContext) bool {
-		return pod.GetPodName() == logContext.GetPrimaryPodName()
-	})
-	if !found {
-		return nil, nil, fmt.Errorf("primary pod %s not found in log context", logContext.GetPrimaryPodName())
-	}
-
-	container, found := lo.Find(pod.GetContainers(), func(c *core.ContainerContext) bool {
-		return c.GetContainerName() == pod.GetPrimaryContainerName()
-	})
-	if !found {
-		return nil, nil, fmt.Errorf("primary container %s not found in pod %s", pod.GetPrimaryContainerName(), pod.GetPodName())
-	}
-
-	return pod, container, nil
 }
